refactor: add commands module
