+++ index.ts
@@ -1,10 +1,6 @@
-import Jimp from 'jimp';
 import { httpServer } from './src/http_server/index.js';
-import robot from 'robotjs';
 import { WebSocketServer } from 'ws';
-import { mouseCommadns } from './src/commands/mouseMoveCommands.js';
-import { drawingCommadns } from './src/commands/drawingCommands.js';
-import { printScreen } from './src/commands/print.js';
+import { commands } from './src/commands/commands.js';
 
 const HTTP_PORT = 3000;
 const HTTP_WS_PORT = 8080;
@@ -20,17 +16,6 @@ wss.on('connection', (ws) => {
   ws.on('message', function incoming(message) {
     const command = message.toString().split(' ')[0];
     const params = message.toString().split(' ').slice(1);
-    if (command === 'mouse_position' ) {
-      const curreentMousePosition = robot.getMousePos();
-      ws.send(`mouse_position ${curreentMousePosition.x},${curreentMousePosition.y}`);
-    } else if (command.includes('mouse_')) {
-      mouseCommadns(command, params);
-      ws.send(command);
-    } else if (command.includes('draw_')) {
-      drawingCommadns(command, params);
-      ws.send(command);
-    } else if (command.includes('prnt_')) {
-      printScreen(ws);
-    }
+    commands(command, ws, params);
   });
 });
